add if for routers

+++ app.js
@@ -1,8 +1,8 @@
 const express = require("express");
 const mongoose = require("mongoose");
 const cors = require("cors");
-const bodyParser = require("body-parser");
 const app = express();
+ _ = require('lodash');
 const Schema = mongoose.Schema;
 
 const taskScheme = new Schema({
@@ -29,26 +29,38 @@ app.get("/allTasks", (req, res) => {
 });
 
 app.post("/createTask", (req, res) => {
-  const task = new Task(req.body);
-  task.save().then((result) => {
-    res.send("Task created");
-  });
+  if (req.body.hasOwnProperty("text") && req.body.hasOwnProperty("isCheck")) {
+    const task = new Task(req.body);
+    task.save().then((result) => {
+      res.send("Task created");
+    });
+  }else {
+    res.status(404).send("Error");
+  }
 });
 
 app.delete("/deleteTask", (req, res) => {
-  Task.deleteOne({ _id: req.query._id }).then((result) => {
-    Task.find().then((result) => {
-      res.send({ data: result });
+  if(req.query._id){
+    Task.deleteOne({ _id: req.query._id }).then((result) => {
+      Task.find().then((result) => {
+        res.send({ data: result });
+      });
     });
-  });
+  }else {
+    res.status(404).send("Error");
+  }
 });
 
 app.patch("/updateTask", (req, res) => {
-  Task.updateOne({ _id: req.query._id }, req.body).then((result) => {
-    Task.find().then((result) => {
-      res.send({ data: result });
+  if (req.body.hasOwnProperty("_id") && (req.body.hasOwnProperty("text") || req.body.hasOwnProperty("isCheck"))){
+    Task.updateOne({ _id: req.query._id }, req.body).then((result) => {
+      Task.find().then((result) => {
+        res.send({ data: result });
+      });
     });
-  });
+} else {
+  res.status(404).send("Error");
+}
 });
 
 app.listen(8000, () => {
